Create LinearSearch.js
Linear search is a very simple search algorithm. In this type of search, a sequential search is made over all items one by one. Every item is checked and if a match is found then that particular item is returned, otherwise the search continues till the end of the data collection.

The time complexity of the above algorithm is O(n). 


Linear search is rarely used practically because other search algorithms such as the binary search algorithm and hash tables allow significantly faster-searching comparison to Linear search.
Improve Linear Search Worst-Case Complexity


if element Found at last  O(n) to O(1)
if element Not found O(n) to O(n/2)